terminando criaturas e implementando data biding

+++ app/src/main/java/com/example/myapplication/model/domain/Creature.kt
@@ -1,8 +1,8 @@
 package com.example.myapplication.model.domain
 
+
 data class Creature(
-    val number: Long,
+    val number: Int,
     val name: String,
     val imageUrl: String
-
 )
